fix(security)!: update deps, drop support for python<=3.6
Remove upper version constraint on `future`. Fixes security issues
GHSA-v3c5-jqr6-7qm8, CVE-2022-40899. (The risk of a new version
breaking things is low, so it is better to allow users to upgrade.)

BREAKING CHANGE: Drop support for python 3.6 and older. The code
will still probably work on older python versions, but it is difficult
to set up test environments for EOL pythons.